src/requirements2.txt
matplotlib==3.3.0
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.1.0
pillow==7.2.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
tensorboard==2.3.0
tensorboardx==2.1
tqdm==4.48.0
tifffile==2019.7.26.2
albumentations==0.4.3
multiprocess==0.70.9
timm==0.3.2
segmentation-models-pytorch==0.1.3
efficientnet-pytorch==0.6.3
xlsxwriter==1.2.6
